+++ shared-libs/ts-modules/start-core/lib/util/index.ts
@@ -3,6 +3,7 @@ export {
   mdnsResolvable,
   filledAddress,
   filterNonLocal,
+  isAddressEnabled,
 } from './filledAddress'
 export type {
   Filter,